refactor(fake-claude): extract script action loop into script.rs module
